use time.Timer.Reset() for flush timeout (#8)
The main benefit is that it's more efficient when handling many idle
connections.  This change eliminates the CPU overhead of an idle network
connection.

In terms of performance, the throughput of BenchmarkWebsocketNetworkBasic
seems to be about the same (maybe slightly better).

In terms of the impact of creating or resetting a timer, the new benchmark
in timer_test.go seems to suggest there's very little cost to resetting
an existing timer (~300 ns per reset + receive, on my laptop), even when
running for many iterations when GC should have kicked in.