Add warning to dial api about response cutoff
It was a bit surprising to find out that only part of the response body was being returned, and I spent time digging around to find the culprit. I think it should at least be documented on the api call.  

Ideally though, it would just propagate the entire response body.