chore: $ go mod -edit go1.13
This commit edited the Go version in go.mod file in order to solve an compile error on the editor on conn.go.

```go
return [4]byte{byte(n), byte(n >> 8), byte(n >> 16), byte(n >> 24)}
```

> invalid operation: signed shift count 8 (untyped int constant) requires go1.13 or later compilerInvalidShiftCount

The unexported function `newMaskKey` was added 4 years ago.
https://github.com/gorilla/websocket/issues/200

However, signed shift counts was restricted before Go 1.13.

> Per the signed shift counts proposal Go 1.13 removes the restriction that a shift count must be unsigned.
https://go.dev/doc/go1.13

It's not affected the behaviour but it's better to mark the Go version is Go 1.13 or later.