Update server.go
After the connection is upgraded from the HTTP protocol to the WebSocket protocol, the network connection cannot be used for HTTP. The net/http server panics when the application attempts to write an HTTP response to a hijacked connection. You should not use netConn to write the response headers. Instead, you should use the http.ResponseWriter to write the response headers to avoid the log "http: response.WriteHeader on hijacked connection from...."

Signed-off-by: samyuktha-asokan <samyuktha.asokan@freshworks.com>